+++ Segnify/Protocols/ForwardedEventsProtocol.swift
@@ -1,5 +1,5 @@
 //
-//  EventsProtocol.swift
+//  ForwardedEventsProtocol.swift
 //  Segnify
 //
 //  Created by Bart Hopster on 30/11/2018.
@@ -18,7 +18,7 @@ import UIKit
 ///
 /// The protocol partly relies on the `UIPageViewControllerDataSource` and `UIPageViewControllerDelegate` protocols.
 /// In this way, the implementations in `PageViewController` aren't at risk.
-public protocol EventsProtocol {
+public protocol ForwardedEventsProtocol {
     
     /// Based on the corresponding `UIPageViewControllerDataSource` protocol, this method returns both the view controller
     /// before the currently viewed view controller, where the user navigated away from, and the latter view controller.
@@ -52,7 +52,7 @@ public protocol EventsProtocol {
 
 // MARK: - Defaults
 
-extension EventsProtocol {
+extension ForwardedEventsProtocol {
     
     /// Provide a default implementation.
     public func pageViewController(_ pageViewController: UIPageViewController,