+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Released on 2018-11-30.
 #### Added
 
 - Added `BannerViewClosure`, for adding one or more subviews to the `bannerView` instance of `PageViewController`.
-- Added `EventsProtocol`.
+- Added `ForwardedEventsProtocol`.
   - Added by [Bart Hopster](https://github.com/barthopster).
 
 ## [1.1.0](https://github.com/nedap/Segnify/releases/tag/1.1.0)